Move JsonC stuff into jsonc package
